microservices_emotions/frontend/requirements.txt
aiofiles==0.8.0
blinker==1.4
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Login==0.6.1
Flask-WTF==1.0.1
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
priority==2.0.0
quart==0.17.0
requests==2.27.1
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wsproto==1.1.0
WTForms==3.0.1
zipp==3.8.0
